requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.11
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
tenacity==9.0.0
python-dateutil==2.9.0.post0
pytest==8.3.3
